+++ src/components/channels/messages/MessagesListItem.vue
@@ -85,6 +85,9 @@
             if (propName.indexOf('timestamp') !== -1) {
               value = date.formatDate(value * 1000, 'DD/MM/YYYY HH:mm:ss')
             }
+            if (propName.indexOf('image.bin.') !== -1) {
+              value = '<binary image>'
+            }
             vals[propName].value = value
           }
           else {