+++ src/components/mqtt/Index.vue
@@ -46,6 +46,7 @@
       'isCustomer',
       'isLoading',
       'isVisibleToolbar',
+      'isNeedSelect',
       'config'
     ],
     data () {